requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
monai[all]>=1.3.0

# Medical imaging
SimpleITK>=2.3.0
nibabel>=5.0.0
pydicom>=2.4.0
rt-utils>=1.2.7
surface-distance @ git+https://github.com/google-deepmind/surface-distance.git

# Data science
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
scikit-image>=0.20.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Experiment tracking
tensorboard>=2.14.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0

# Notebook
jupyter>=1.0.0
ipykernel>=6.25.0
ipywidgets>=8.0.0
